refactor: migrate loading secton to new component
